+++ src/lib/dhcp/option_vendor_class.cc
@@ -71,7 +71,7 @@ OptionVendorClass::unpack(OptionBufferConstIter begin,
 
     // Start reading opaque data.
     size_t offset = 0;
-    while (offset < std::distance(begin, end)) {
+    while (offset < static_cast<size_t>(std::distance(begin, end))) {
         // Parse a tuple.
         OpaqueDataTuple tuple(OptionDataTypeUtil::getTupleLenFieldType(getUniverse()),
                               begin + offset, end);
